:rocket: Finished the method to check for wins!
Moves are populated into a 2d array If all bad request checks return false,

Then the board and player that made the move is passed into a method that checks for wins.

Win method checks if all 3 in a row combinations contains its own ID if that comes back as true this constitutes as a win.